backend/requirements.txt
Flask
Flask_SQLAlchemy
Flask_Migrate
Flask-JWT-Extended
pymysql
python-dotenv
cryptography
celery
redis
flask_mail
flasagger